Adicionado o comunicação com os outros serviços no serviço order

+++ services/order/src/main/java/com/filipe/ecommerce/order/OrderController.java
@@ -5,6 +5,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.List;
+
 @RestController
 @RequestMapping( "/api/v1/orders" )
 @RequiredArgsConstructor
@@ -13,17 +15,17 @@ public class OrderController {
     private final OrderService service;
 
     @PostMapping
-    public ResponseEntity<Integer> createOrder( @RequestBody @Valid OrderRequest request ) {
-        return ResponseEntity.ok( service.createOrder(request));
+    public ResponseEntity<Long> createOrder( @RequestBody @Valid OrderRequest request ) {
+        return ResponseEntity.ok( service.createOrder( request ) );
     }
 
     @GetMapping
     public ResponseEntity<List<OrderResponse>> findAll() {
-        return ResponseEntity.ok( service.findAllOrders());
+        return ResponseEntity.ok( service.findAllOrders() );
     }
 
     @GetMapping("/{orderId}")
-    public ResponseEntity<OrderResponse> findById( @PathVariable("orderId") Integer orderId ) {
+    public ResponseEntity<OrderResponse> findById( @PathVariable("orderId") Long orderId ) {
         return ResponseEntity.ok( service.findById( orderId ) );
     }
 }
